refactor(api): rename verified-code-route
